Add mlp.py and modify common directory construction
Squashed commit of the following:

commit 2ffcb4057a32e48d2831ee3e47d0bfe8ca49afeb
Author: khkim <kh.kim@medipixel.io>
Date:   Fri Feb 1 18:08:00 2019 +0900

    Add license

commit 537a20e148d2413f80e86032db82a843164637a8
Author: khkim <kh.kim@medipixel.io>
Date:   Fri Feb 1 16:25:55 2019 +0900

    Modify common structure

commit fab8d8a50608b0a9a081d50e39d39cc8927902d8
Author: khkim <kh.kim@medipixel.io>
Date:   Fri Feb 1 15:47:49 2019 +0900

    Add mlp.py and modify bc-ddpg

commit 8eebb317175a3eeda3305bfdad5916a5b3a6a059
Author: khkim <kh.kim@medipixel.io>
Date:   Fri Feb 1 11:35:55 2019 +0900

    Add mlp model
