+++ algorithms/common/buffer/segment_tree.py
