+++ algorithms/common/helper_functions.py
