+++ algorithms/per/ddpg_agent.py
@@ -17,12 +17,10 @@ import torch
 import torch.optim as optim
 import wandb
 
-import algorithms.common.utils.helper_functions as common_utils
+import algorithms.common.helper_functions as common_utils
 from algorithms.common.abstract.agent import AbstractAgent
-from algorithms.common.noise.ou_noise import OUNoise
-from algorithms.common.replaybuffer.priortized_replay_buffer import (
-    PrioritizedReplayBuffer,
-)
+from algorithms.common.buffer.priortized_replay_buffer import PrioritizedReplayBuffer
+from algorithms.common.noise import OUNoise
 from algorithms.ddpg.model import Actor, Critic
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
