+++ algorithms/td3/agent.py
@@ -17,10 +17,10 @@ import torch.nn.functional as F
 import torch.optim as optim
 import wandb
 
-import algorithms.common.utils.helper_functions as common_utils
+import algorithms.common.helper_functions as common_utils
 from algorithms.common.abstract.agent import AbstractAgent
-from algorithms.common.noise.gaussian_noise import GaussianNoise
-from algorithms.common.replaybuffer.replay_buffer import ReplayBuffer
+from algorithms.common.buffer.replay_buffer import ReplayBuffer
+from algorithms.common.noise import GaussianNoise
 from algorithms.td3.model import Actor, Critic
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
